+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Peripheral proxy now generate peripheral names in same way as CC:T modems.
   They will be unique, but numbers will raise significantly.
 
+### Changed
+
+- `solidCoolantTemp` now exists for powah, actually same as `coolantTemp`
+
 ### Fixed
 
 - Peripheral names are always sorted